requirements.txt
requests>=2.25.0
python-dotenv>=0.19.0
tqdm>=4.64.0
colorama>=0.4.4